+++ R/characterize_edit_site.R
@@ -42,64 +42,64 @@ characterize_edit_site <- function(edit_site_row) {
     'description' = edit_site_row
   )
 
-  # 1. Filter BAM
+  # Filter BAM
   df_bam_filtered <- filter_bam_alignments(
     df_bam = df_bam,
     specimen = edit_site$description$specimen,
     chromosome = edit_site$description$chromosome,
     edit_site_position = edit_site$description$position,
     edit_site_strand = edit_site$description$strand,
-    allowed_bases_within_gRNA = run_params$cut_site_start_distance_within_gRNA,
-    allowed_bases_outside_gRNA = run_params$cut_site_start_distance_outside_gRNA
+    allowed_aln_start_cut_site_PAM_distal = run_params$allowed_aln_start_cut_site_PAM_distal,
+    allowed_aln_start_cut_site_PAM_proximal = run_params$allowed_aln_start_cut_site_PAM_proximal
   )
 
   if (nrow(df_bam_filtered) == 0) {
-    return(NULL) # Equivalent of 'next'
+    edit_site$fail <- 'no rows after bam filtering'
+    saveRDS(edit_site, file.path(run_params$analysis_output, 'edit_sites', paste0('edit_site_id_', edit_site$description$unique_edit_site_id, '.rds')))
+    return(NULL)
   }
 
-  # 2. Process Alignments
+  # Process Alignments
   df_aln_pos <- aln_to_base_and_position_table(df_bam_filtered = df_bam_filtered)
 
-  df_aln_pos_filtered <- calculate_base_percentages(
-    df_aln_pos = df_aln_pos,
-    edit_site_strand = edit_site$description$strand,
-    edit_site_position = edit_site$description$position,
-    bases_from_cut_site = run_params$end_distance_from_cut_site
-  )
 
-  # 3. Handle Coordinates & Reference
+  # Standardize coordinates, add reference
   edit_site$protospacer <- cut_site_to_protospacer_coordinates(
     cut_site = edit_site$description$position,
     strand = edit_site$description$strand,
-    end_distance_from_cut_site = run_params$end_distance_from_cut_site,
-    expected_cut_distance_from_pam = run_params$expected_cut_distance_from_pam,
+    protospacer_distance_from_cut_PAM_distal = run_params$expected_PAM_distal_cut_distance,
+    protospacer_distance_from_cut_PAM_proximal = run_params$expected_PAM_proximal_cut_distance,
     pam_length = 3
   )
 
-  if (edit_site$description$strand == '+') {
-    start_pos <- edit_site$protospacer$protospacer_end
-    stop_pos <- edit_site$protospacer$PAM_start
-  } else {
-    start_pos <- edit_site$protospacer$PAM_start
-    stop_pos <- edit_site$protospacer$protospacer_end
-  }
+  absolute_start_stop_coords <- protospacer_coords_to_absolute_start_stop_coords(
+    protospacer_coords = edit_site$protospacer,
+    strand = edit_site$description$strand
+    )
 
   edit_site$genomic_reference <- extract_reference_bases(
     genomic_sequence = genome_sequence,
     chromosome = edit_site$description$chromosome,
-    start = start_pos,
-    stop = stop_pos
+    start = absolute_start_stop_coords$start,
+    stop = absolute_start_stop_coords$stop
   )
 
-  # 4. Significance Testing
+  # Calculate percentage of each base
+  df_aln_pos_filtered <- calculate_base_percentages(
+    df_aln_pos = df_aln_pos,
+    start = absolute_start_stop_coords$start,
+    stop = absolute_start_stop_coords$stop
+  )
+
+  # Significance Testing
   df_base_percentages_with_reference <- binomial_prop_edit_test(
     df_base_counts = df_aln_pos_filtered,
-    p_value_threshold = run_params$binomial_p_value_threshold,
+    null_proportion = run_params$null_proportion,
     alternative_ = run_params$binomial_direction
   ) %>%
     dplyr::left_join(edit_site$genomic_reference, by = 'position')
 
-  # 5. Strand Correction
+  # Strand Correction
   if (edit_site$description$strand == '-') {
     df_base_percentages_with_reference$reference_base <- as.character(unlist(sapply(df_base_percentages_with_reference$reference_base, base_complement)))
     df_base_percentages_with_reference$base <- as.character(unlist(sapply(df_base_percentages_with_reference$base, base_complement)))
@@ -109,10 +109,11 @@ characterize_edit_site <- function(edit_site_row) {
     df_edit_significance = df_base_percentages_with_reference,
     editable_base = run_params$editable_base,
     expected_edit = run_params$expected_edit,
-    pvalue_threshold = run_params$binomial_p_value_threshold
+    pvalue_threshold = run_params$binomial_p_value_threshold,
+    adj_method = run_params$p_value_adj_method
   )
 
-  # 6. Final Formatting
+  # Final Formatting
   edit_site$base_composition <- df_base_percentages_with_reference %>%
     dplyr::select(position, position_depth, base, base_count, percentage, reference_base)
 
@@ -123,7 +124,7 @@ characterize_edit_site <- function(edit_site_row) {
     bases = edit_site$genomic_reference$reference_base
   )
 
-  # 7. Save and Clean up
+  # Save and Clean up
   saveRDS(edit_site, file.path(run_params$analysis_output, 'edit_sites', paste0('edit_site_id_', edit_site$description$unique_edit_site_id, '.rds')))
 
   return(edit_site$description$unique_edit_site_id) # Return ID to track completion