+++ R/quantify_edits.R
@@ -13,17 +13,19 @@
 #' @param output_directory Character. Directory where output data will be saved.
 #' @param analysis_name Character. Name for this analysis (default: `'quantify_edits'`).
 #' @param abundance_cutoff Numeric. Minimum read count for including an edit site (default: `3`).
-#' @param expected_cut_distance_from_pam Numeric. Distance from PAM where the cut site is expected (default: `3`).
-#' @param end_distance_from_cut_site Numeric. Maximum distance from the cut site to consider bases (default: `20`).
-#' @param cut_site_start_distance_within_gRNA Numeric. Distance within the gRNA to start considering cut sites (default: `3`).
-#' @param cut_site_start_distance_outside_gRNA Numeric. Distance outside the gRNA to start considering cut sites (default: `3`).
+#' @param expected_PAM_proximal_cut_distance Numeric. PAM proximal distance of cut site to start of PAM (default: `3`).
+#' @param expected_PAM_distal_cut_distance Numeric. Maximum PAM-distal distance from the cut site to consider bases (default: `17`).
+#' @param allowed_aln_start_cut_site_PAM_distal Numeric. Distance within the gRNA to start considering cut sites (default: `3`).
+#' @param allowed_aln_start_cut_site_PAM_proximal Numeric. Distance outside the gRNA to start considering cut sites (default: `3`).
 #' @param reference_genome_path Character. Path to a FASTA file used to retrieve reference genome sequence (default: `''`).
 #' @param editable_base Character. The original base that is targeted for editing (default: `'A'`).
 #' @param expected_edit Character. The expected base after editing (default: `'G'`).
+#' @param null_proportion Numeric. The expected background/error rate under the null hypothesis (default: `0.001`).
 #' @param binomial_p_value_threshold Numeric. P-value threshold for determining significance using a binomial test (default: `0.05`).
 #' @param binomial_direction Character. Direction of the binomial test: `'greater'`, `'less'`, or `'two-sided'` (default: `'greater'`).
+#' @param p_value_adj_method Character. Must match accepted method from p.adjust(). (default: `'BH'`).
 #' @param n_processors Numeric. Number of CPU cores to use for parallel processing of BAM files (default: `4`).
-#' @param n_edit_site_processors Numeric. Number of CPU cores to use for parallel processing of edit sites with significant levels of base editing.
+#' @param n_edit_site_processors Numeric. Number of CPU cores to use for parallel processing of edit sites with significant levels of base editing (default: `3`).
 #' @param overwrite Logical. Whether to overwrite existing analysis output (default: `TRUE`).
 #'
 #' @return This function does not return an object. It writes the following to `output_directory/analysis_name`:
@@ -43,10 +45,10 @@
 #'   output_directory = "path/to/output_directory",
 #'   analysis_name = "my_analysis",
 #'   abundance_cutoff = 5,
-#'   expected_cut_distance_from_pam = 3,
-#'   end_distance_from_cut_site = 25,
-#'   cut_site_start_distance_within_gRNA = 4,
-#'   cut_site_start_distance_outside_gRNA = 4,
+#'   expected_PAM_proximal_cut_distance = 3,
+#'   expected_PAM_distal_cut_distance = 17,
+#'   allowed_aln_start_cut_site_PAM_distal = 4,
+#'   allowed_aln_start_cut_site_PAM_proximal = 4,
 #'   reference_genome_path = "path/to/genome.fasta",
 #'   editable_base = "C",
 #'   expected_edit = "T",
@@ -58,130 +60,137 @@
 #' )
 #' }
 #' @export
-
 quantify_edits <- function(
 		base_directory,
 		output_directory,
 		analysis_name = 'quantify_edits',
 		abundance_cutoff = 3,
-		expected_cut_distance_from_pam = 3,
-		end_distance_from_cut_site = 20,
-		cut_site_start_distance_within_gRNA = 3,
-		cut_site_start_distance_outside_gRNA = 3,
+		expected_PAM_proximal_cut_distance = 3,
+		expected_PAM_distal_cut_distance = 17,
+		allowed_aln_start_cut_site_PAM_distal = 3,
+		allowed_aln_start_cut_site_PAM_proximal = 3,
 		reference_genome_path = '',
 		editable_base = 'A',
 		expected_edit = 'G',
+		null_proportion = 0.001,
 		binomial_p_value_threshold = 0.05,
-		binomial_direction = 'greater',
+		binomial_direction = c('greater', 'less', 'two.sided'),
+		p_value_adj_method = 'BH',
 		n_processors = 4,
 		n_edit_site_processors = 3,
 		overwrite = TRUE
 ) {
 
-  library(BSgenome)
-  library(logr)
+	library(BSgenome)
+	library(logr)
 
-  run_params <- data.frame(
-  	'base_directory' = base_directory,
-  	'output_directory' = output_directory,
-  	'analysis_name' = analysis_name,
-  	'abundance_cutoff' = abundance_cutoff,
-  	'expected_cut_distance_from_pam' = expected_cut_distance_from_pam,
-  	'end_distance_from_cut_site' = end_distance_from_cut_site,
-  	'cut_site_start_distance_within_gRNA' = cut_site_start_distance_within_gRNA,
-  	'cut_site_start_distance_outside_gRNA' = cut_site_start_distance_outside_gRNA,
-  	'reference_genome_path' = reference_genome_path,
-  	'editable_base' = editable_base,
-  	'expected_edit' = expected_edit,
-  	'binomial_p_value_threshold' = binomial_p_value_threshold,
-  	'binomial_direction' = binomial_direction,
-  	'n_processors' = n_processors,
-  	'n_edit_site_processors' = n_edit_site_processors,
-  	'overwrite' = overwrite
-  )
+	# Resolve default arguments
+	binomial_direction <- match.arg(binomial_direction)
 
-  ## Troubleshooting start ##
-  # run_params <- data.frame(
-  #   'base_directory' = '/data/BEiGUIDE/data-raw/230705_MN01490_0144_A000H5KVFN',
-  #   'output_directory' = '/data/BEiGUIDE/tests',
-  #   'analysis_name' = 'quantify_edits',
-  #   'abundance_cutoff' = 5,
+	run_params <- data.frame(
+			'base_directory' = base_directory,
+			'output_directory' = output_directory,
+			'analysis_name' = analysis_name,
+			'abundance_cutoff' = abundance_cutoff,
+			'expected_PAM_proximal_cut_distance' = expected_PAM_proximal_cut_distance,
+			'expected_PAM_distal_cut_distance' = expected_PAM_distal_cut_distance,
+			'allowed_aln_start_cut_site_PAM_distal' = allowed_aln_start_cut_site_PAM_distal,
+			'allowed_aln_start_cut_site_PAM_proximal' = allowed_aln_start_cut_site_PAM_proximal,
+			'reference_genome_path' = reference_genome_path,
+			'editable_base' = editable_base,
+			'expected_edit' = expected_edit,
+			'null_proportion' = null_proportion,
+			'binomial_p_value_threshold' = binomial_p_value_threshold,
+			'binomial_direction' = binomial_direction,
+			'p_value_adj_method' = p_value_adj_method,
+			'n_processors' = n_processors,
+			'n_edit_site_processors' = n_edit_site_processors,
+			'overwrite' = overwrite
+	)
 
-  #   'expected_cut_distance_from_pam' = 3,
-  #   'end_distance_from_cut_site' = 20,
-  #   'cut_site_start_distance_within_gRNA' = 3,
-  #   'cut_site_start_distance_outside_gRNA' = 3,
-  #   'reference_genome_path' = '/data/iGUIDE/genomes/hg38.fasta',
+	## Troubleshooting start ##
+	# run_params <- data.frame(
+	#    'base_directory' = '/data/friederike_herbst_nowrouzi_project/projects/base_editor_ptprc_project/data/processed/0-iGUIDE_runs/params1_260529_MN01490_0363_A000HCJLNT',
+	#    'output_directory' = '/data/friederike_herbst_nowrouzi_project/projects/base_editor_ptprc_project/data/processed/7-replicate_comparison/1-run_beiguide',
+	#    'analysis_name' = 'quantify_edits',
+	#    'abundance_cutoff' = 1,
+	#    'expected_PAM_proximal_cut_distance' = 3,
+	#    'expected_PAM_distal_cut_distance' = 17,
+	#    'allowed_aln_start_cut_site_PAM_distal' = 3,
+	#    'allowed_aln_start_cut_site_PAM_proximal' = 3,
+	#    'reference_genome_path' = '/data/iGUIDE/genomes/hg38.fasta',
+	#    'editable_base' = 'A',
+	#    'expected_edit' = 'G',
+	#    'null_proportion' = 0.001,
+	#    'binomial_p_value_threshold' = 0.05,
+	#    'binomial_direction' = 'greater',
+	#    'p_value_adj_method' = 'BH',
+	#    'n_processors' = 24,
+	#    'n_edit_site_processors' = 3,
+	#    'overwrite' = TRUE
+	# )
+	## Troubleshooting End ##
 
-  #   'editable_base' = 'A',
-  #   'expected_edit' = 'G',
 
-  #   'binomial_p_value_threshold' = 0.15,
-  #   'binomial_direction' = 'greater',
+	run_params$analysis_output <- file.path(run_params$output_directory, run_params$analysis_name)
+	manage_run_directory(run_params = run_params)
+	saveRDS(run_params, file.path(run_params$analysis_output, 'run_parameters.rds'))
+	write.csv(run_params, file.path(run_params$analysis_output, 'run_parameters.csv'), row.names = F)
 
-  #   'n_processors' = 2,
-  #   'overwrite' = TRUE
-  # )
-  ## Troubleshooting End ##
+	dir.create(file.path(run_params$analysis_output, 'edit_sites'))
+	dir.create(file.path(run_params$analysis_output, 'plots'))
 
-  run_params$analysis_output <- file.path(run_params$output_directory, run_params$analysis_name)
-  manage_run_directory(run_params = run_params)
-  saveRDS(run_params, file.path(run_params$analysis_output, 'run_parameters.rds'))
-  write.csv(run_params, file.path(run_params$analysis_output, 'run_parameters.csv'), row.names = F)
+	logr::log_open(file_name = file.path(run_params$analysis_output, 'BEiGUIDE_quantify_edits'), logdir = FALSE)
 
-  dir.create(file.path(run_params$analysis_output, 'edit_sites'))
-  dir.create(file.path(run_params$analysis_output, 'plots'))
+	logr::log_print(run_params)
 
-  logr::log_open(file_name = file.path(run_params$analysis_output, 'BEiGUIDE_quantify_edits'), logdir = FALSE)
+	df_ft_data <- pull_ft_data_tables(base_directory = run_params$base_directory)
 
-  logr::log_print(run_params)
+	df_annotations <- pull_combo_overview_table(base_directory = run_params$base_directory)
 
-  df_ft_data <- pull_ft_data_tables(base_directory = run_params$base_directory)
+	df_edit_sites <- make_edit_site_table(
+		ft_data_table = df_ft_data,
+		spec_info_combo_overview_table = df_annotations,
+		abundance_cutoff = run_params$abundance_cutoff)
 
-  df_annotations <- pull_combo_overview_table(base_directory = run_params$base_directory)
+	saveRDS(df_edit_sites, file.path(run_params$analysis_output, 'edit_sites_overview.rds'))
+	write.csv(df_edit_sites, file.path(run_params$analysis_output, 'edit_sites_overview.csv'), row.names = F)
 
-  df_edit_sites <- make_edit_site_table(
-  	ft_data_table = df_ft_data,
-  	spec_info_combo_overview_table = df_annotations,
-  	abundance_cutoff = run_params$abundance_cutoff)
+	bam_files <- list_bam_files(base_directory = run_params$base_directory)
 
-  saveRDS(df_edit_sites, file.path(run_params$analysis_output, 'edit_sites_overview.rds'))
-  write.csv(df_edit_sites, file.path(run_params$analysis_output, 'edit_sites_overview.csv'), row.names = F)
+	logr::log_print(bam_files)
 
-  bam_files <- list_bam_files(base_directory = run_params$base_directory)
+	df_bam <- parallel_bam_file_list_to_table(
+		bam_file_list = bam_files,
+		scan_param_what_list = c('qname', 'rname', 'strand', 'pos', 'qwidth', 'seq', 'cigar', 'flag'),
+		number_of_cpu = run_params$n_processors
+	)
 
-  logr::log_print(bam_files)
+	# Filter once for speed.
+	df_bam <- df_bam %>%
+		dplyr::filter(
+			!stringr::str_detect(cigar, 'I'),
+			!stringr::str_detect(cigar, 'S'),
+			!stringr::str_detect(cigar, 'D')
+		)
 
-  df_bam <- parallel_bam_file_list_to_table(
-  	bam_file_list = bam_files,
-  	scan_param_what_list = c('qname', 'rname', 'strand', 'pos', 'qwidth', 'seq', 'cigar', 'flag'),
-  	number_of_cpu = run_params$n_processors
-  )
+	genome_sequence <- load_reference_genome(run_params$reference_genome_path)
 
-  # Filter once for speed.
-  df_bam <- df_bam %>%
-    dplyr::filter(
-      !stringr::str_detect(cigar, 'I'),
-      !stringr::str_detect(cigar, 'S'),
-      !stringr::str_detect(cigar, 'D')
-    )
+	df_edit_sites_pass_abundance <- df_edit_sites %>%
+		dplyr::filter(pass_abundance_filter == T)
 
-  genome_sequence <- load_reference_genome(run_params$reference_genome_path)
+	# edit_sites to process
+	edit_site_list <- split(
+		df_edit_sites_pass_abundance,
+		1:nrow(df_edit_sites_pass_abundance)
+	)
 
-  df_edit_sites_pass_abundance <- df_edit_sites %>%
-  	dplyr::filter(pass_abundance_filter == T)
+	# A shortcut to load all params and
+	base::assign("df_bam", df_bam, envir = .GlobalEnv)
+	base::assign("run_params", run_params, envir = .GlobalEnv)
+	base::assign("genome_sequence", genome_sequence, envir = .GlobalEnv)
 
-  # edit_sites to process
-  edit_site_list <- split(
-  	df_edit_sites_pass_abundance,
-  	1:nrow(df_edit_sites_pass_abundance)
-  )
-
-  base::assign("df_bam", df_bam, envir = .GlobalEnv)
-  base::assign("run_params", run_params, envir = .GlobalEnv)
-  base::assign("genome_sequence", genome_sequence, envir = .GlobalEnv)
-
-  logr::log_print(paste("Starting parallel characterization of", length(edit_site_list), "sites using", run_params$n_edit_site_processors, "cores."))
+	logr::log_print(paste("Starting parallel characterization of", length(edit_site_list), "sites using", run_params$n_edit_site_processors, "cores."))
 
 	results <- parallel::mclapply(
 		edit_site_list,