+++ tests/testthat/test-cut_site_to_protospacer_coordinates.R
@@ -2,8 +2,8 @@ test_that("cut_site_to_protospacer_coordinates works for + strand", {
   result <- cut_site_to_protospacer_coordinates(
     cut_site = 100,
     strand = '+',
-    end_distance_from_cut_site = 20,
-    expected_cut_distance_from_pam = 3,
+    protospacer_distance_from_cut_PAM_distal = 20,
+    protospacer_distance_from_cut_PAM_proximal = 3,
     pam_length = 3
   )
 
@@ -13,15 +13,15 @@ test_that("cut_site_to_protospacer_coordinates works for + strand", {
   # + strand logic
   expect_equal(result$protospacer_start, 103)
   expect_equal(result$PAM_start, 106)
-  expect_equal(result$protospacer_end, 80)
+  expect_equal(result$protospacer_end, 81)
 })
 
 test_that("cut_site_to_protospacer_coordinates works for - strand", {
   result <- cut_site_to_protospacer_coordinates(
     cut_site = 100,
     strand = '-',
-    end_distance_from_cut_site = 20,
-    expected_cut_distance_from_pam = 3,
+    protospacer_distance_from_cut_PAM_distal = 20,
+    protospacer_distance_from_cut_PAM_proximal = 3,
     pam_length = 3
   )
 
@@ -31,10 +31,10 @@ test_that("cut_site_to_protospacer_coordinates works for - strand", {
   # - strand logic
   expect_equal(result$protospacer_start, 97)
   expect_equal(result$PAM_start, 94)
-  expect_equal(result$protospacer_end, 120)
+  expect_equal(result$protospacer_end, 119)
 })
 
 test_that("cut_site is coerced to numeric", {
-  result <- cut_site_to_protospacer_coordinates("100", "+")
+  result <- cut_site_to_protospacer_coordinates(cut_site = "100",strand = "+")
   expect_equal(result$protospacer_start, 103)
 })